Fix support for vertical bar identifiers.
The parser was not stepping to the next token after parsing the
identifier. This caused the closing vertical bar to be `duplicated'
within the parser. For example, `(|define| a 2)' was being parsed as
`(|define| |a 2)', which would cause `a 2)' to be parsed as an
Identifier, resulting in a parse error when there was no closing
vertical bar found.